backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.8.5
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
notion-client>=2.2.1
python-multipart>=0.0.6
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
